Create function_overloading.cpp
Polymorphism
Polymorphism allows objects to be treated as instances of their parent class rather than their actual class. It enables one interface to be used for a general class of actions. There are two types of polymorphism in C++: compile-time (or static) polymorphism and run-time (or dynamic) polymorphism.

Compile-time Polymorphism (Function Overloading and Operator Overloading)